pyproject.toml
[tool.poetry]
name = "python-async-lambda"
version = "0.1.0"
description = ""
authors = ["Ville Kärkkäinen <ville.karkkainen@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aioboto3 = "11.2.0"
aws-xray-sdk = "^2.12.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
